+++ recipes/server.rb
@@ -16,7 +16,7 @@ include_recipe "logrotate"
 include_recipe "rabbitmq" if node['logstash']['install_rabbitmq']
 
 if node['logstash']['install_zeromq']
-  include_recipe "yumrepo::zeromq" if platform_family?("rhel")
+  include_recipe 'logstash::zero_mq_repo'
   node['logstash']['zeromq_packages'].each {|p| package p }
 end
 
@@ -185,6 +185,26 @@ services.each do |type|
       else
         Chef::Log.fatal("Please set node['logstash']['server']['init_method'] to 'runit' for #{node['platform_version']}")
       end
+
+    elsif platform_family? "fedora" and node["platform_version"] >= "15"
+      execute "reload-systemd" do
+        command "systemctl --system daemon-reload"
+        action :nothing
+      end
+
+      template "/etc/systemd/system/logstash_server.service" do
+        source "logstash_server.service.erb"
+        owner "root" and group 'root' and mode 0755
+        notifies :run, "execute[reload-systemd]", :immediately
+        notifies :restart, "service[logstash_server]", :delayed
+      end
+
+      service "logstash_server" do
+        service_name "logstash_server.service"
+        provider Chef::Provider::Service::Systemd
+        action [ :enable, :start ]
+      end
+
     elsif platform_family? "rhel","fedora"
       template "/etc/init.d/logstash_#{type}" do
         source "init.logstash_#{type}.erb"